+++ src/test/java/test/TestFormatter.java
@@ -3,10 +3,9 @@ package test;
 import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.StringReader;
-import java.io.StringWriter;
 
-import be.baur.sda.SDA;
 import be.baur.sda.DataNode;
+import be.baur.sda.SDA;
 import be.baur.sda.serialization.SDAFormatter;
 
 public final class TestFormatter {
@@ -20,10 +19,10 @@ public final class TestFormatter {
 			return s;
 		});
 		
-		StringWriter out = new StringWriter();
+
         final DataNode node = SDA.parse(new StringReader("node\"1\"{node2{empty1\"\"empty2{}empty\"3\"{}}}"));
-		formatter4.format(out, node);
-		t.ts1("S01", out.toString(), 
+		String str = formatter4.format(node);
+		t.ts1("S01", str, 
 				"node \"1\" {\n" + 
 				"    node2 {\n" + 
 				"        empty1 \"\"\n" +
@@ -32,11 +31,10 @@ public final class TestFormatter {
 				"    }\n" + 
 				"}\n");
 		
-		out = new StringWriter();
 		InputStream in = TestFormatter.class.getResourceAsStream("/addressbook.sda");
 		final DataNode book = SDA.parse(new InputStreamReader(in,"UTF-8"));
-		formatter.format(out, book);
-		t.ts1("S02", out.toString(), "addressbook {\n" + 
+		str = formatter.format(book);
+		t.ts1("S02", str, "addressbook {\n" + 
 				"	contact \"1\" {\n" + 
 				"		firstname \"Alice\"\n" + 
 				"		phonenumber \"06-11111111\"\n" + 
@@ -55,17 +53,16 @@ public final class TestFormatter {
 //		UnitTestPerformance<Node> perf = new UnitTestPerformance<Node>(
 //			n -> { n.toString(); }   // test toString
 //		);
-//		perf.run("\nP01", node, 25000, 25);
+//		perf.run("\nP01", book, 25000, 25);
 
-//		final StringWriter sw = new StringWriter();
-//		UnitTestPerformance<Node> perf = new UnitTestPerformance<Node>(
-//		n -> { try {
-//			formatter.format(sw, book);   // test formatter
-//		} catch (IOException e) {
-//			e.printStackTrace();
-//		} }
-//	);
-//	perf.run("\nP02", node, 25000, 25);
-		
+//		UnitTestPerformance<Node> perf = new UnitTestPerformance<Node>(n -> {
+//			try {
+//				formatter.format(book); // test formatter
+//			} catch (IOException e) {
+//				e.printStackTrace();
+//			}
+//		});
+//		perf.run("\nP02", node, 25000, 25);
+
 	}
 }